[2116] Ядро / Разное: Добавлена обертка для фикса [2115] (780c6d949523) которая позволяет создавать ауру с определенным количеством чарджей и модчарджей.

+++ src/server/game/Spells/Auras/SpellAuras.cpp
@@ -853,7 +853,7 @@ void Aura::SetCharges(uint8 charges)
     SetNeedClientUpdateForTargets();
 }
 
-void Aura::SetModCharges(uint8 charges)
+void Aura::SetModCharges(int16 charges)
 {
     SetCharges(charges);
     for (uint8 i = EFFECT_0; i < EFFECT_2; ++i)
